+++ src/types/mod.rs
@@ -5,6 +5,8 @@ pub(crate) use std::any::Any;
 pub(crate) use std::sync::Arc;
 pub(crate) use std::borrow::Cow;
 pub(crate) use std::sync::atomic;
+pub(crate) use std::fmt::Display;
+pub(crate) use std::fmt::Formatter;
 pub(crate) use std::future::Future;
 pub(crate) use std::net::SocketAddr;
 pub(crate) use std::time::SystemTime;