+++ src/types/scheme.rs
@@ -1,4 +1,5 @@
 //! HTTP Scheme
+use crate::types::*;
 
 /// HTTP Scheme
 #[derive(Debug, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
@@ -8,4 +9,13 @@ pub enum Scheme {
 
     /// HTTPS
     HTTPS,
+}
+
+impl Display for Scheme {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Scheme::HTTP => write!(f, "http"),
+            Scheme::HTTPS => write!(f, "https"),
+        }
+    }
 }