+++ HISTORY.md
@@ -13,7 +13,7 @@
 - Domain filtering
 - Subdomain filtering
 - Subrouter support
-- Response builder
+- Varied responses
 - Multiple listeners
 
 ## [Todo]